web_app.py
import streamlit as st
from PIL import Image
# Import your existing functions here

st.title("Traid Card Scanner")

# Web-based file uploader
uploaded_file = st.file_uploader("Upload a photo of your cards", type=['jpg', 'jpeg', 'png'])

if uploaded_file is not None:
    # Display the uploaded image
    image = Image.open(uploaded_file)
    st.image(image, caption='Uploaded Image', use_column_width=True)
    
    if st.button('Scan Cards'):
        with st.spinner('AI is identifying cards...'):
            # Call your identify_multiple_cards function using uploaded_file
            # results = identify_multiple_cards(uploaded_file)
            st.success("Scan Complete! Check your Supabase dashboard.")