requirements.txt
streamlit
Pillow
requests
supabase
huggingface_hub
